include eslint as peer and regular ("peer with default")
As instructed at https://yarnpkg.com/advanced/rulebook,

> The preferred one is to list the dependency (in Next.js's
case, webpack) as _both a regular dependency and a peer
dependency_. Yarn will interpret this pattern as "peer
dependency with a default", meaning that your users will be
able to take ownership of the Webpack package if they need
to, while still giving the package manager the ability to
emit a warning if the provided version is incompatible with
the one your package expects.

Emphasis in the original.